Remove the universal input API module
It was really just a very complicated way of pre-building input queries.
The problem is that the way you should do this in Rust involves a great
deal of generic nesting, which greatly inflates the typename. This
will push users to generate the queries each frame, which means the API
just becomes unnecessary overhead to the normal input API.

Resolves #154