Make every drawable texturable
To make this convenient and performant, the best option is to expose
much more of the internals to the Drawable objects. This means that a
Mesh is the best option, which means that exposing Mesh from Window also
makes a good deal of sense.

RenderTarget no longer is necessary, because a generic draw function can
just take a mesh and use Drawable's trait function to draw to it. Window
methods are still the cleanest way to draw, and can be used in
situations where drawing directly to the Window makes the most sense.