Create new immi APIs that take view into account
The only way to correctly use immi with the old API would be to write
the UI to a Mesh, then project all the vertices in the Mesh to match the
view of the Window, then drawing it to the window. The new API allows
arbitrary views to be used, and also makes idiomatic use of immi easier.