Change the get_region API to screenshot()
Screenshot is a more discoverable name and the provided region would be
in the OpenGL Viewport space, not in the space the user draws are
projected from. No amount of warnings and documentation would make that
obvious.

Additionally, expecting users to do something useful with the byte
array, un-flip the image (thanks to OpenGL's bottom-left origin), and
probably use the image crate themselves seems unreasonable when the
image crate can just be used.