Add bounds checks to Font::render
Rusttype can produce negative minimum bounds as well as sizes greater
than the specified font size.

In the former case, we would get an overflow:

    thread 'main' panicked at 'attempt to add with overflow'

In the latter an out-of-bounds check trying to write to the `pixels`
Vec:

    thread 'main' panicked at 'index out of bounds: the len is 20736 but the index is 20884'

It's not clear what exactly is causing this, but it can be triggered by
a combination of a specific text and font.

This commit adds bounds checks to handle both cases. Similar checks can
be found in the rusttype example:

https://gitlab.redox-os.org/redox-os/rusttype/blob/1846aae84d1a60b493bf3d55fd41a76367f7f787/examples/simple.rs#L60-61