Fix Immi mouse click handling
Immi's `image_button` was not returning `Interaction::Clicked` even when the
button was clicked.  The reason was that the `ImmiStatus` quicksilver generated
relied on the mouse button to be `ButtonState::Pressed` but due to an earlier
`clear_temporary_states` call in the application, it was always
`ButtonState::Held` instead.

This checks for `state.is_down()` instead, which is true for both `Pressed` and
`Held` cases and makes Immi return the `Clicked` interaction.