Fix broken formatting in documentation (#504)
* Fix missing linebreaks in documentation

Also reflows much of the text to fit within the recommended
100-column width and fixes a couple of tiny grammatical errors.

* Fix broken logo link

The logo at the top of the main doc page was linking to a `logo.svg` that was
not present in the documentation directory, thus causing it to always show the
alt text. I have changed it to a link to the logo on GitHub.