Remove all the old code from 0.3.x
quicksilver-next is shaping up to be nearly a complete rewrite, because
of the change to use ecosystem support for web instead of manual support
for web. This means a good deal of the old support abstractions need to
go.

Additionally, I want to make a move away from magic towards easy-to-use
orthogonal abstractions. Instead of a View and a ResizeStrategy applied
under the hood, simple functions to set the view matrix based on resize
events can serve more use cases with less potential for surprisng
behavior.