Release the first alpha of version 0.4 (#553)
* Start on a complete overhaul to use new libs

* Sketch an outline of the new API

The lifecycle will just re-export Blinds, whereas the new graphics API
looks more like Java2D than it does Quicksilver 0.3 or the new
underlying graphics lib, Golem.

This could totally change, but it solves a few problems (like lack of
clarity) from the old API.

* Add some projection API

* Update to the new Golem prepared draw APIs

This should make the draw API faster, and also removes the soundness
hole from the previous version

* Allow different element types other than Triangles

This means line rendering in the new graphics stack will actually
render lines, not triangularized "lines."

Also improve the graphics handling of state changes

* Add a function for a stroke path

* Ensure all indices passed to GL are valid

This is part of the safety contract of the Golem draw functions

* Add methods to draw circles

* Fix a bug where images wouldn't draw

The image binding wasn't matching up with the uniform value, because 0
has been reserved by Golem

* Add image tinting and subimages to the draw stack

Currently there isn't a way to draw an image at a given size, but
there also isn't a general transform stack.

* Rename geometry_mode to geom_mode_changes

* Add a second function to the transform stack

It allows applying transforms on the GPU without changing the
projection, for all the rotation anyone could need

* Add logging to the 'run' implementation

* Add the 'static/' folder feature

Also add some log support, which should make downstream logging much
easier. It's an optional feature but it's super convenient for dealing
with different logging implementations on desktop and both web
platforms.

* Remove all the old code from 0.3.x

quicksilver-next is shaping up to be nearly a complete rewrite, because
of the change to use ecosystem support for web instead of manual support
for web. This means a good deal of the old support abstractions need to
go.

Additionally, I want to make a move away from magic towards easy-to-use
orthogonal abstractions. Instead of a View and a ResizeStrategy applied
under the hood, simple functions to set the view matrix based on resize
events can serve more use cases with less potential for surprisng
behavior.

* Run a cargot fmt pass

* Ignore a clippy lint about literals

* Add Github CI for the next branch

Also remove the travis config

* Test examples to cover doctests as well

* Start writing docs for the new version

* WIP: Re-introduce the old geometry module

* Fix a row/column major screwup

* Run a formatting pass

* Link from the result to the QuicksilverError

* Add docs to the error

* Improve and document the graphics module

Remove DrawGroup, as it was just an implementation detail of Mesh

Move the transform to the CPU instead of the GPU, because it's likely to
change more often.

* Remove the old lifecycle file

* Add more lifecycle documentation

* Make gamepad support an optional feature

* Remove #[non_exhaustive] for now

* Update docs for the new alpha version

* Update project documentation

* Bump the version

* Add the new version to the changelog