Create the 'input' module to replace 'lifecycle' (#604)
First, the 'lifecycle' concept is less important than in previous
versions. There is just a single 'run' function that takes a Settings
instance and a closure, instead of a whole mess of behind-the-scenes
work. Most of the 'lifecycle' module (with the exception of a few Window
functions) is just about receiving input from the user. The new name
makes this more clear.

Additionally, there were some ergonomic issues with re-exporting blinds
structures rather than wrapping them in new code. Primarily, this
required understanding "what is the mint crate" and "why do I care" and
"how do I use it." End users of Quicksilver shouldn't have to concern
themselves with this, so these parts of the blinds API have been wrapped
over.

The Settings struct also needed to be decoupled from blinds. There are
more settings than just window management (for example, disable
searching for 'static/' assets) so there are now Quicksilver-specific
Settings.

The old module and its members have been left around for convenience,
because the alpha frequently requires updating for new features or
critical bugfixes. However, they are strongly deprecated and will be
subject to removal before beta.

Also, re-export Graphics and Input at the top level. They are needed
in literally every Quicksilver application, because you need to name the
types to run 'app.'